+++ readers/filters/filters.py
@@ -7,39 +7,51 @@ from .value_getters import ValueGetter
 from .operations import Operation
 
 class Filter(object):
-    def __init__(self, operation, op1, op2, transform=None):
+    def __init__(self, operator, op1, op2, **kwargs):
         """
-        - op1_getter: jq-like string with the query that fetches the key, which
+        - operator: operation name that will be used.
+        - op1: jq-like string with the query that fetches the key, which
             will be used as op1.
-        - operation: operation name that will be used.
         - op2: value for the second operand of `operation`.
         - transform: if present apply this function to op1 after getting it
         """
-        self.op1_getter = ValueGetter(op1, transform)
-        self.operation = Operation(operation)
-        self.op2 = op2
+        try:
+            self.op1 = ValueGetter(op1, **kwargs)
+        except ValueError as err:
+            try:
+                self.op1 = Query(**op1)
+            except TypeError as err:
+                raise err
+        try:
+            self.op2 = ValueGetter(op2, **kwargs)
+        except ValueError as err:
+            try:
+                self.op2 = Query(**op2)
+            except TypeError as err:
+                self.op2 = op2
+        self.operator = Operation(operator)
 
     def __str__(self):
-        return " - ".join([str(self.op1_getter), str(self.operation), str(self.op2)])
+        return '(' + ' - '.join([str(self.op1), str(self.operator), str(self.op2)]) + ')'
 
     def __repr__(self):
         return self.__str__()
 
     def __call__(self, js):
         """
-        Return whether the JSON object passes the filter.
+        Walks the tree and returns the boolean
         """
-        op1 = self.op1_getter(js)
-        return self.operation(op1, self.op2)
+        op1 = self.op1(js)
+        try:
+            op2 = self.op2(js)
+        except TypeError as err:
+            op2 = self.op2
+
+        return self.operator(op1, op2)
 
     @classmethod
     def fromConfig(cls, config):
         """
-        Create a filter from a dict with specifications.
+        Create a query from a dict with specifications.
         """
-        op1_getter = config['op1']
-        transform = config.get('transform')
-        operation = config['operator']
-        op2 = config['op2']
-
-        return cls(op1_getter, operation, op2, transform)
+        return cls(**config)
